+++ app/src/main/java/com/fibelatti/photowidget/widget/PrepareCurrentPhotoUseCase.kt
@@ -73,7 +73,7 @@ class PrepareCurrentPhotoUseCase @Inject constructor(
             is PhotoWidgetBorder.None -> null
             is PhotoWidgetBorder.Color -> "#${photoWidget.border.colorHex}".toColorInt()
             is PhotoWidgetBorder.Dynamic -> context.getDynamicAttributeColor(
-                com.google.android.material.R.attr.colorPrimary,
+                com.google.android.material.R.attr.colorPrimaryInverse,
             )
         }
         val borderPercent = photoWidget.border.getBorderPercent()