+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Fixed
+
+* Fix which dynamic color is applied to the widget border
+
 ## [v1.24.0] - 2025-02-13
 [v1.24.0]: https://github.com/fibelatti/photo-widget/releases/tag/v1.24.0